svg/pyproject.toml
[project]
name = "svg-workflow"
version = "2.0.0"
description = "LangGraph 文档拆分 + 智能绘图工作流"
readme = "REFACTORING.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "python-docx>=1.1.0",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]

[tool.uv]
python-preference = "only-system"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
